+++ lib/src/analyzer/analyzer.dart
@@ -1,5 +1,5 @@
 // ignore_for_file: avoid_dynamic_calls
-// Copyright (c) 2021 - 2023 Buijs Software
+// Copyright (c) 2021 - 2025 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -114,8 +114,8 @@ AnalysisResult analyze({
   }
 
   final result = file!.path.contains(metadataMarkerPrefix)
-      ? file.parseMetadata
-      : file.parseDataClass;
+      ? file.parseMetadata()
+      : file.parseDataClass();
 
   if (pathToOutputFolder != null) {
     if (!Directory(pathToOutputFolder).existsSync()) {
@@ -131,7 +131,7 @@ AnalysisResult analyze({
 /// {@category analyzer}
 extension FileAnalyzer on File {
   /// Use [JsonVisitor] to collect Metadata from dart class.
-  AnalysisResult get parseDataClass {
+  AnalysisResult parseDataClass() {
     final visitor = JsonVisitor();
     if (!_visitFile(visitor)) {
       return const AnalysisResult(parent: null);
@@ -154,7 +154,7 @@ extension FileAnalyzer on File {
         childrenEnumTypes: types.whereType<EnumType>().toSet());
 
     if (parentType is CustomType) {
-      return result.normalizeParentTypeMembers;
+      return result.normalizeParentTypeMembers();
     } else {
       return result;
     }
@@ -177,7 +177,7 @@ extension FileAnalyzer on File {
   /// JSON decode current file and return [CustomType].
   ///
   /// The JSON is expected to contain metadata for a single data class.
-  AnalysisResult get parseMetadata {
+  AnalysisResult parseMetadata() {
     final enumTypes = <EnumType>[];
     final customTypes = <CustomType>[];
     final json = readAsStringSync().jsonDecode;
@@ -194,25 +194,29 @@ extension FileAnalyzer on File {
           final nullable = object["nullable"] as bool;
 
           final memberType = type.toAbstractType(nullable: nullable);
-
-          if (memberType is CustomType || memberType is EnumType) {
-            final debugFile = parent.resolve(
-                "$metadataMarkerPrefix${memberType.className.toLowerCase()}.json");
-            if (debugFile.existsSync()) {
-              final result = debugFile.parseMetadata;
-              final parentOrNull = result.parent;
-              if (parentOrNull is CustomType) {
-                customTypes.add(parentOrNull);
+          switch (memberType) {
+            case EnumType() || CustomType() || NonCanonicalType():
+              final debugFile = parent.resolve(
+                  "$metadataMarkerPrefix${memberType.className.toLowerCase()}.json");
+              if (debugFile.existsSync()) {
+                final result = debugFile.parseMetadata();
+                final parentOrNull = result.parent;
+                if (parentOrNull is CustomType) {
+                  customTypes.add(parentOrNull);
+                }
+                if (parentOrNull is EnumType) {
+                  enumTypes.add(parentOrNull);
+                }
+                customTypes.addAll(result.childrenCustomTypes);
+                enumTypes.addAll(result.childrenEnumTypes);
+              } else {
+                "Found ${memberType.runtimeType} but no source (Does not exist: ${debugFile.path})"
+                    .log();
               }
-              if (parentOrNull is EnumType) {
-                enumTypes.add(parentOrNull);
-              }
-              customTypes.addAll(result.childrenCustomTypes);
-              enumTypes.addAll(result.childrenEnumTypes);
-            } else {
-              "Found ${memberType.runtimeType} but no source (Does not exist: ${debugFile.path})"
-                  .log();
-            }
+            case StandardType():
+              ;
+            case UndeterminedAsDynamic():
+              ;
           }
 
           members.add(
@@ -230,7 +234,7 @@ extension FileAnalyzer on File {
           ),
           childrenCustomTypes: customTypes.toSet(),
           childrenEnumTypes: enumTypes.toSet(),
-        ).normalizeParentTypeMembers;
+        ).normalizeParentTypeMembers();
       }
     }
 
@@ -270,6 +274,7 @@ extension FileAnalyzer on File {
       } 
     ]
   }""");
+
     "Example of EnumType metadata JSON file:".log(context: """
     {
         "className": "MyResponse",
@@ -364,7 +369,7 @@ extension AbstractTypeSerializer on AbstractType {
 }
 
 extension on AnalysisResult {
-  AnalysisResult get normalizeParentTypeMembers {
+  AnalysisResult normalizeParentTypeMembers() {
     if (parent == null) {
       return this;
     }