+++ lib/src/decoder/array.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2023 Buijs Software
+// Copyright (c) 2021 - 2025 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -328,7 +328,9 @@ extension JsonArrayDecoder on String {
   /// Example:
   ///
   /// Given a JSON Array:
+  /// ```dart
   /// [[[ "hello", "goodbye" ], ["Sam", "Hanna"]]]
+  /// ```
   ///
   /// Would result in the following Map:
   /// ".0.0.0.0" : "hello"
@@ -408,7 +410,9 @@ extension ListDecoder on List {
   /// Example:
   ///
   /// Given a List:
+  /// ```dart
   /// [[[ "hello", "goodbye" ], ["Sam", "Hanna"]]]
+  /// ```
   ///
   /// Would result in the following Map:
   /// ".0.0.0.0" : "hello"
@@ -446,7 +450,8 @@ class ListOpeningBracketToken extends _Token {
           depth: currentDepth + 1,
           size: currentSize,
           value: currentValue,
-          key: keygen.nextKey(output.keys.toList(), ListOpeningBracketToken),
+          key: keygen.nextKey(
+              output.keys.toList(), _ListTokenType.listOpeningBracket),
         );
 }
 
@@ -469,7 +474,8 @@ class ListClosingBracketToken extends _Token {
             currentKey: keygen.currentKey,
             output: output,
           ),
-          key: keygen.nextKey(output.keys.toList(), ListClosingBracketToken),
+          key: keygen.nextKey(
+              output.keys.toList(), _ListTokenType.listClosingBracket),
         );
 }
 
@@ -518,7 +524,8 @@ class ListValueSeparatorToken extends _Token {
             currentKey: keygen.currentKey,
             output: output,
           ),
-          key: keygen.nextKey(output.keys.toList(), ListValueSeparatorToken),
+          key: keygen.nextKey(
+              output.keys.toList(), _ListTokenType.listValueSeparator),
         );
 
   static Map<int, int> _size(Map<int, int> currentSize, int depth) {
@@ -668,6 +675,12 @@ int _setObjectValueAndReturnRemainder({
   }
 }
 
+enum _ListTokenType {
+  listOpeningBracket,
+  listClosingBracket,
+  listValueSeparator,
+}
+
 abstract class _Token {
   _Token({
     required this.size,
@@ -759,15 +772,15 @@ class ArrayDecodingKeyGenerator {
   String currentKey;
 
   /// Determine the next key value depending on the last encountered token.
-  String nextKey(List<String> keys, Type token) {
+  String nextKey(List<String> keys, _ListTokenType token) {
     switch (token) {
-      case ListOpeningBracketToken:
+      case _ListTokenType.listOpeningBracket:
         currentKey = "$currentKey.0";
         break;
-      case ListClosingBracketToken:
+      case _ListTokenType.listClosingBracket:
         currentKey = currentKey.substring(0, currentKey.lastIndexOf("."));
         break;
-      case ListValueSeparatorToken:
+      case _ListTokenType.listValueSeparator:
         incrementWidth;
     }
 
@@ -816,13 +829,13 @@ List buildListStructure<T>(List<List<int>> positions, {List<T>? valueList}) {
   );
 }
 
-/// Get a List<T> with 0 or more parent lists.
+/// Get a List&lt;T&gt; with 0 or more parent lists.
 ///
 /// Specify [depth] to added one or more parent Lists.
 ///
 /// Example:
-/// [depth] 0 = List<T>
-/// [depth] 2 = List<List<List<T>>>.
+/// [depth] 0 = List&lt;T&gt;
+/// [depth] 2 = List&lt;List&lt;List&lt;T&gt;&gt;&gt;.
 List getNestedList<T>({
   required int depth,
   required List<T> valueList,
@@ -837,13 +850,13 @@ List getNestedList<T>({
 /// Add another List arround the given List and keep the it strongly typed.
 List<List<T>> _addParent<T>(List<T> list) => [list];
 
-/// Get a List<String> with 0 or more parent lists.
+/// Get a List&lt;String&gt; with 0 or more parent lists.
 ///
 /// Specify [depth] to added one or more parent Lists.
 ///
 /// Example:
-/// [depth] 0 = List<String>
-/// [depth] 2 = List<List<List<String>>>.
+/// [depth] 0 = List&lt;String&gt;
+/// [depth] 2 = List&lt;List&lt;List&lt;String&gt;&gt;&gt;.
 List getNestedStringList(int depth) =>
     getNestedList<String>(depth: depth, valueList: <String>[]);
 
@@ -851,13 +864,13 @@ List getNestedStringList(int depth) =>
 List getNestedNullableStringList(int depth) =>
     getNestedList<String?>(depth: depth, valueList: <String?>[]);
 
-/// Get a List<int> with 0 or more parent lists.
+/// Get a List&lt;int&gt; with 0 or more parent lists.
 ///
 /// Specify [depth] to added one or more parent Lists.
 ///
 /// Example:
-/// [depth] 0 = List<int>
-/// [depth] 2 = List<List<List<int>>>.
+/// [depth] 0 = List&lt;int&gt;
+/// [depth] 2 = List&lt;List&lt;List&lt;int&gt;&gt;&gt;.
 List getNestedIntList(int depth) =>
     getNestedList<int>(depth: depth, valueList: <int>[]);
 
@@ -865,13 +878,13 @@ List getNestedIntList(int depth) =>
 List getNestedNullableIntList(int depth) =>
     getNestedList<int?>(depth: depth, valueList: <int?>[]);
 
-/// Get a List<double> with 0 or more parent lists.
+/// Get a List&lt;double&gt; with 0 or more parent lists.
 ///
 /// Specify [depth] to added one or more parent Lists.
 ///
 /// Example:
-/// [depth] 0 = List<double>
-/// [depth] 2 = List<List<List<double>>>.
+/// [depth] 0 = List&lt;double&gt;
+/// [depth] 2 = List&lt;List&lt;List&lt;double&gt;&gt;&gt;.
 List getNestedDoubleList(int depth) =>
     getNestedList<double>(depth: depth, valueList: <double>[]);
 
@@ -879,13 +892,13 @@ List getNestedDoubleList(int depth) =>
 List getNestedNullableDoubleList(int depth) =>
     getNestedList<double?>(depth: depth, valueList: <double?>[]);
 
-/// Get a List<bool> with 0 or more parent lists.
+/// Get a List&lt;bool&gt; with 0 or more parent lists.
 ///
 /// Specify [depth] to added one or more parent Lists.
 ///
 /// Example:
-/// [depth] 0 = List<bool>
-/// [depth] 2 = List<List<List<bool>>>.
+/// [depth] 0 = List&lt;bool&gt;
+/// [depth] 2 = List&lt;List&lt;List&lt;bool&gt;&gt;&gt;.
 List getNestedBoolList(int depth) =>
     getNestedList<bool>(depth: depth, valueList: <bool>[]);