+++ lib/src/decoder/utils.dart
@@ -1,4 +1,4 @@
-// Copyright (c) 2021 - 2023 Buijs Software
+// Copyright (c) 2021 - 2025 Buijs Software
 //
 // Permission is hereby granted, free of charge, to any person obtaining a copy
 // of this software and associated documentation files (the "Software"), to deal
@@ -20,7 +20,6 @@
 
 import "../ast/ast.dart";
 import "../common/common.dart";
-import "../converters/undetermined.dart";
 
 /// Find matching AbstractType for String value.
 ///
@@ -62,8 +61,8 @@ extension AbstractTypeFromString on String {
       return type;
     }
 
-    final customType = withoutPostfix._toCustomTypeOrNull;
-
+    final customType =
+        withoutPostfix._toNonCanonicalTypeOrNull(nullable ?? false);
     if (customType != null) {
       return customType;
     }
@@ -75,9 +74,9 @@ extension AbstractTypeFromString on String {
     throw SquintException("Unable to determine type: '$this'");
   }
 
-  CustomType? get _toCustomTypeOrNull {
+  NonCanonicalType? _toNonCanonicalTypeOrNull(bool nullable) {
     final hasMatch = _customClassNameRegex.hasMatch(this);
-    return hasMatch ? CustomType(className: this, members: []) : null;
+    return hasMatch ? NonCanonicalType(this) : null;
   }
 }
 
@@ -139,14 +138,14 @@ AbstractType? _mapType({
 
 /// Regex to match a List from literal value:
 ///
-/// List<...>
+/// List&lt;...&gt;
 ///
 /// {@category decoder}
 final _listRegex = RegExp(r"""^(List)(<(.+?)>|)$""");
 
 /// Regex to match a Map from literal value:
 ///
-/// Map<...,...>
+/// Map&lt;...,...&gt;
 ///
 /// {@category decoder}
 final mapRegex = RegExp(r"""^(Map)(<(.+?),(.+?)>|)$""");